Rename node on board as cell
Prepend to a linked list adds a node into the linked list taking
the value you give it (in this case an object called Node), and then
also adding a next property to create an object called Node pointing to
another node. The problem here is there are two conceptual nodes which
are for different things. This becomes confused in the code. Therefore
rename the board nodes with coordinates as cells.