Fix children inherit parents but parents seem to lose their own parents
I put a breakpoint at (2,1) to track assignment of own parent
and assignment of it's children cells to it as a parent and seeing where
its parent went. Turns out its parent dissapeared and that was only
because instead of referencing frontCell when it was (2,1) which had all
the information, I referenced grid[i][j].value which mapped to the x and
y of front cell. Technically it should reference the same cell. But it
turns out javascript references don't work like that. Editing the parent
property of the node's cell put into the queue is not the same as
editing the grid node object's cell value directly. References in JS.